Added ability to review records, view different results
