+++ package.json
@@ -5,17 +5,22 @@
 	"license": "MIT",
 	"dependencies": {
 		"@svgr/webpack": "5.5.0",
+		"aws-amplify": "^4.0.3",
 		"html-react-parser": "^1.2.3",
 		"jstoxml": "^2.0.3",
 		"next": "^10.0.5",
+		"prop-types": "^15.7.2",
 		"raw-loader": "^4.0.2",
 		"react": "^17.0.1",
+		"react-avatar": "^3.10.0",
 		"react-dom": "^17.0.1",
 		"react-redux": "^7.2.3",
 		"redux": "^4.0.5",
 		"redux-thunk": "^2.3.0",
 		"sass": "^1.32.13",
+		"short-uuid": "^4.2.0",
 		"typescript": "~4.2.4",
+		"uuid": "^8.3.2",
 		"web3-utils": "^1.3.6",
 		"workbox-core": "^6.1.2",
 		"workbox-expiration": "^6.0.2",
@@ -47,7 +52,15 @@
 		"eslint": "eslint ./src"
 	},
 	"browserslist": {
-		"production": [ ">0.2%", "not dead", "not op_mini all" ],
-		"development": [ "last 1 chrome version", "last 1 firefox version", "last 1 safari version" ]
+		"production": [
+			">0.2%",
+			"not dead",
+			"not op_mini all"
+		],
+		"development": [
+			"last 1 chrome version",
+			"last 1 firefox version",
+			"last 1 safari version"
+		]
 	}
 }
