+++ src/components/record-card/RecordCard.module.scss
@@ -1,0 +1,33 @@
+.card {
+    display: block;
+    width: 100%;
+    box-sizing: border-box;
+    padding: 1rem 2rem;
+    margin: 0 0 1rem;
+    border: var(--border);
+    border-radius: var(--border-radius);
+    background-color: var(--primary-block-color);
+    color: var(--primary-text-color);
+    text-align: center;
+    cursor: pointer;
+
+    span,
+    svg {
+        --icon-size: 19px;
+    }
+
+    p,
+    figcaption {
+        font-size: 16px;
+        font-weight: 400;
+        text-align: center;
+    }
+
+    p {
+        margin: 1rem 0;
+    }
+
+    figcaption {
+        font-weight: 500;
+    }
+}
