+++ src/components/record-card/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RecordCard.component';
