+++ src/pages/stakes/index.tsx
