+++ src/pages/stakes/stake/[id].tsx
@@ -1,0 +1,1 @@
+export { default } from 'routes/stakes/record/[id]';
